Update Endpoint Config
