+++ backend/functions/src/index.ts
@@ -4,17 +4,19 @@ import { initializeApp } from "firebase-admin/app";
 initializeApp();
 
 // Exports (Keep Alphabetized)
-export { adminDeleteRecord } from "./api/adminDeleteRecord";
-export { clientSignIn } from "./api/clientSignIn";
-export { createManagerProfile } from "./api/createManagerProfile";
-export { getResident } from "./api/getResident";
 export { helloWorld } from "./api/helloWorld";
-export { onAcceptCall } from "./api/onAcceptCall";
-export { onCancelCall } from "./api/onCancelCall";
-export { onManagerLogin } from "./api/onManagerLogin";
-export { onManagerLogout } from "./api/onManagerLogout";
-export { onManagerTokenRefresh } from "./api/onManagerTokenRefresh";
-export { onRejectCall } from "./api/onRejectCall";
-export { onRequestCall } from "./api/onRequestCall";
-export { onResidentLogout } from "./api/onResidentLogout";
-export { onResidentTokenRefresh } from "./api/onResidentTokenRefresh";
+
+// Legacy
+// export { adminDeleteRecord } from "./api/legacy/adminDeleteRecord";
+// export { clientSignIn } from "./api/legacy/clientSignIn";
+// export { createManagerProfile } from "./api/legacy/createManagerProfile";
+// export { getResident } from "./api/legacy/getResident";
+// export { onAcceptCall } from "./api/legacy/onAcceptCall";
+// export { onCancelCall } from "./api/legacy/onCancelCall";
+// export { onManagerLogin } from "./api/legacy/onManagerLogin";
+// export { onManagerLogout } from "./api/legacy/onManagerLogout";
+// export { onManagerTokenRefresh } from "./api/legacy/onManagerTokenRefresh";
+// export { onRejectCall } from "./api/legacy/onRejectCall";
+// export { onRequestCall } from "./api/legacy/onRequestCall";
+// export { onResidentLogout } from "./api/legacy/onResidentLogout";
+// export { onResidentTokenRefresh } from "./api/legacy/onResidentTokenRefresh";
